Refactored directory structure and added barrel files
