+++ lib/features/authentication/authentication.dart
@@ -1,0 +1,3 @@
+export 'authentication_service.dart';
+export 'biometric_auth_page.dart';
+export 'pin_auth_page.dart';
