+++ lib/features/chart/chart.dart
@@ -1,5 +1,5 @@
-export 'abstract_chart_provider.dart';
 export 'base_chart_widget.dart';
 export 'base_heatmap_widget.dart';
-export 'counter_chart_provider.dart';
-export 'timer_chart_provider.dart';
+export 'provider/abstract_chart_provider.dart';
+export 'provider/counter_chart_provider.dart';
+export 'provider/timer_chart_provider.dart';
