+++ lib/features/onboarding/onboarding.dart
@@ -1,1 +1,1 @@
-export 'onboarding_screen.dart';
+export 'onboarding_page.dart';
