+++ lib/features/splash/splash.dart
@@ -1,1 +1,1 @@
-export 'splash_screen.dart';
+export 'splash_page.dart';
